WebEvolver/cognitive_kernel_v2/backend/requirements.txt
aiohttp==3.9.5
aioprocessing==2.0.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
distro==1.9.0
dnspython==2.6.1
docx2txt==0.8
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
faiss-cpu==1.7.4
fastapi==0.111.0
fastapi-cli==0.0.3
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
passlib==1.7.4
pdfminer.six==20231228
pillow==10.3.0
pip==23.0.1
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
rsa==4.9
setuptools==58.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wheel==0.43.0
yarl==1.9.4
bs4
sanic